Back-End :: Addressing Build-related issues

+++ src/main/java/com/devops/aquarium/web/controller/ActivityController.java
@@ -2,6 +2,8 @@ package com.devops.aquarium.web.controller;
 
 import com.devops.aquarium.dao.ActivityDao;
 import com.devops.aquarium.model.Activity;
+import com.devops.aquarium.web.exceptions.BRException;
+import com.devops.aquarium.web.exceptions.ISException;
 import com.devops.aquarium.web.exceptions.IdNotFoundException;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -13,6 +15,7 @@ import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import javax.validation.Valid;
 import java.net.URI;
+import java.util.List;
 
 @Api( description="API for, inter allia, CRUD-related operations on activities.")
 @RestController
@@ -24,14 +27,17 @@ public class ActivityController {
 
     @RequestMapping(value="all/activity", method= RequestMethod.GET)
     public Iterable<Activity> findAllActivities() {
-        return activityDao.findAll();
+
+        List<Activity> activity = activityDao.findAll();
+        if(activity == null) throw new ISException("I_S");//Customized 500
+        return activity;
     }
 
     @ApiOperation(value = "Send back a given activity")
     @GetMapping(value = "all/activity/{id}")
     public Activity findActivityById(@PathVariable int id) {
         Activity activity=activityDao.findById(id);
-        if(activity==null) throw new IdNotFoundException("Wrong Id"); //Customized exception throwing
+        if(activity==null) throw new IdNotFoundException("Wrong Id"); //Customized exception encompassing '404 NOT FOUND' case
         return activity;
     }
 
@@ -53,6 +59,9 @@ public class ActivityController {
 
     @DeleteMapping (value = "delete/activity/{id}")
     public void deleteActivity(@PathVariable int id) {
+
+        Activity activity=activityDao.findById(id);
+        if(activity==null) throw new IdNotFoundException("Wrong Id"); //Customized exception encompassing '404 NOT FOUND' case
         activityDao.deleteById(id);
     }
 
@@ -60,6 +69,11 @@ public class ActivityController {
             consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE},
             produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
     public Activity updateActivity(@Valid @RequestBody Activity activity) {
-        return activityDao.save(activity);
+
+        Activity activityA=activityDao.findById(activity.getId());
+        if (activityA==null) throw new IdNotFoundException("Wrong Id"); //Customized exception encompassing '404 NOT FOUND' case
+        Activity activityT=activityDao.save(activity);
+        if (activityT==null) throw new BRException("BR"); //Customized 400
+        return activityT;
     }
 }
